+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # CHANGELOG:
 
+### 2021-10-08
+* Modified "main.py"
+* Recreated "menu.py" from 2021-09-22 as "menu_sample.py"
+    * Code imported and modified from previous "menu.py"
+    * Added a changelog inside file
+    * Fixed a function breaking indentation error
+    * Fixed docstrings, blank lines to improve readability
+
+### 2021-10-07
+* Modified "main.py"
+
+### 2021-10-06
+* Modified "main.py"
+
 ### 2021-10-05
 * Modified "main.py"